auto-adjudication/requirements.txt
aws-cdk-lib>=2.208.0
constructs>=10.3.0

